edited readme's and postprocessing quest cleanup-zip script

+++ nucluster/cleanup_and_zip_simFiles.py
@@ -24,11 +24,19 @@ def parse_args():
         default="Local"
     )
     parser.add_argument(
+        "-zd",
         "--zip_dir",
         action='store_true',
         help="If specified archives the simulation folder",
     )
     parser.add_argument(
+        "-dt",
+        "--del_trajectories",
+        action='store_true',
+        help="""If specified deletes single trajectories and simulation emodl files"""
+    )
+    parser.add_argument(
+        "-dd",
         "--del_dir",
         action='store_true',
         help="""If specified deletes the simulation folder, 
@@ -51,13 +59,14 @@ def make_archive(source, destination):
     shutil.move('%s.%s' % (name, format), destination)
 
 
-def cleanup_and_zip_simFiles(sim_output_path, zip_dir, delete_from_dir):
+def cleanup_and_zip_simFiles(sim_output_path, del_trajectories, zip_dir, del_dir):
     """Delete single trajectories and simulation files"""
-    shutil.rmtree(os.path.join(sim_output_path,'trajectories'), ignore_errors=True)
-    shutil.rmtree(os.path.join(sim_output_path,'simulations'), ignore_errors=True)
-    if args.zip_dir:
+    if del_trajectories:
+        shutil.rmtree(os.path.join(sim_output_path,'trajectories'), ignore_errors=True)
+        shutil.rmtree(os.path.join(sim_output_path,'simulations'), ignore_errors=True)
+    if zip_dir:
         make_archive(sim_output_path, os.path.join(wdir, 'simulation_output', f'{exp_name}_zip.zip'))
-    if args.delete_from_dir:
+    if del_dir:
         shutil.rmtree(os.path.join(sim_output_path), ignore_errors=True)
 
 if __name__ == '__main__':
@@ -69,6 +78,9 @@ if __name__ == '__main__':
 
     exp_names = [x for x in os.listdir(os.path.join(wdir, 'simulation_output')) if stem in x]
     for exp_name in exp_names:
-        print("Zipping files for " + exp_name)
+        print("Cleanup files for " + exp_name)
         sim_output_path = os.path.join(wdir, 'simulation_output', exp_name)
-        cleanup_and_zip_simFiles(sim_output_path, zip_dir=args.zip_dir, del_dir=args.del_dir)
+        cleanup_and_zip_simFiles(sim_output_path,
+                                 del_trajectories=args.del_trajectories,
+                                 zip_dir=args.zip_dir,
+                                 del_dir=args.del_dir)
